Optimizations

+++ NetworkingPrimitivesCore.Tests/IPv4NetworkTests.cs
@@ -87,7 +87,7 @@ public class IPv4NetworkTests
     public void IPv4Network_Subnet_Test(string networkString, int prefix, int index, string subnetString)
     {
         var network = IPv4Network.Parse(networkString);
-        var subnet = network.Subnet(prefix, index);
+        var subnet = network.Subnet((byte)prefix, index);
         Assert.AreEqual(subnetString, subnet.ToString());
     }
 
@@ -102,7 +102,7 @@ public class IPv4NetworkTests
     public void IPv4Network_Supernet_Test(string networkString, int prefix, string supernetString)
     {
         var network = IPv4Network.Parse(networkString);
-        var supernet = network.Supernet(prefix);
+        var supernet = network.Supernet((byte)prefix);
         Assert.AreEqual(supernetString, supernet.ToString());
     }
